requirements.txt
Django==4.2.6
mariadb==1.1.7
mysqlclient==2.2.0
python-dotenv==1.0.0